added bit util operations, minor file structure improvements
